refactor: dto 하위에 projection 폴더 추가
